extracted the builder to its separate class as it can lead to conflicts in groovy
