+++ src/main/java/tkcy/tktech/api/unification/properties/MaterialPropertiesAddition.java
@@ -23,7 +23,7 @@ public class MaterialPropertiesAddition {
         ToxicMaterialProperty.addToxicMaterialProperty(Materials.Benzene);
         ToxicMaterialProperty.addToxicMaterialProperty(Materials.CoalGas);
         CorrosiveMaterialProperty.addCorrosiveMaterialProperty(TkTechMaterials.HydrogenFluoride);
-        PhysicalProperties.Builder builder = new PhysicalProperties.Builder().autoIgnitionTemperature(503).bp(348)
+        PhysicalPropertiesBuilder builder = new PhysicalPropertiesBuilder().autoIgnitionTemperature(503).bp(348)
                 .mp(176)
                 .flameTemperature(2120);
         PhysicalProperties.addPhysicalMaterialProperty(TkTechMaterials.HydrogenFluoride, builder);
