+++ src/main/java/tkcy/tktech/api/unification/properties/TkTechMaterialPropertyKeys.java
@@ -13,7 +13,8 @@ public class TkTechMaterialPropertyKeys {
     public static final PropertyKey<CorrosiveMaterialProperty> CORROSIVE = new PropertyKey<>("corrosive",
             CorrosiveMaterialProperty.class);
 
-    public static final PropertyKey<PhysicalProperties> PHYSICAL_PROPERTIES = new PropertyKey<>("physical-properties",
+    public static final PropertyKey<PhysicalProperties> PHYSICAL_PROPERTIES = new PropertyKey<>(
+            "physical_properties",
             PhysicalProperties.class);
 
     static {
